services/robustness-orchestrator/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Numerical & Scientific
numpy>=1.24.0
scipy>=1.10.0

# AI & Reporting
tenacity>=8.2.0
fpdf2>=2.7.0

# Kimi AI (Moonshot AI) - OpenAI-compatible API
openai>=1.0.0

# Caching & Messaging
redis>=4.5.0

# Supabase (Demo Access & User Profiles)
supabase>=2.3.0

# Observability
structlog>=23.0.0
prometheus-client>=0.17.0

# Billing
stripe>=5.0.0

# AI Token Counting
tiktoken>=0.5.0

# PDF Charts
matplotlib>=3.7.0

# Production Server
gunicorn>=21.0.0
celery>=5.3.0

# Utilities
httpx>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# PDF Fonts (for Unicode support)
# DejaVu Sans TTF should be placed in fonts/ directory
